More c++11/14 stuff like autos

+++ runq.cc
@@ -699,7 +699,7 @@ main (int argc, char** argv)
         default:
             cerr << endl
                  << "*** ERROR: Unhandled RunQ exception ("
-                 << error.error
+                 << static_cast <int> (error.error)
                  << "), aborting ..."
                  << endl;
             if (error.info != 0)
